Multitasking and dynamic thread registration
Multitasking and dynamic thread registration.

In this example we have given every thread its own stack. Now threads can have function calls, parameters and local variables etc. Another important thing in this example is that the creation of threads is now dynamic. The thread registration code initializes the PCB, and adds it to the linked list so that the scheduler will give it a turn.