Character Input using DOS services
Character Input using DOS services.

The DOS input buffer has a special format where the first byte stores the maximum characters buffer can hold, the second byte holds the number of characters actually read on return, and the following space is used for the actual characters read. In this example of reading a string with service 1 and displaying it with service 9.