Another Multitasking TSR caller
Another Multitasking TSR caller.

Multitasking Kernel Client: As an example, the multitasking kernel client is modified to execute a new instance of the command prompt after running all three threads. This demonstrates how the INT 21, AH = 4Bh service can be used to load and execute a new program (in this case, the command prompt) after completing a task.