Transfer data back and forth using serial port
A program using serial port to transfer data back amd forth.

This is the example where two computers are connected using a serial cable made just as described above. The program is to be run on both computers. After that whatever is typed on one computer appears on the screen of the other.