32-bit Programming and going into protected mode
32-bit Programming and going into protected mode.

The aim of this example is to give a falvour of 32bit programming. We have written the printstr function for read and for protected mode. The availability of larger registers and flexible addressing rules allows writing a much comprehensive version of the code. Also offsets to parameters and default widths change.